refactored: split offers stores by role
